feat: Add master link for overall data sharing
Introduces a new "Master Link" feature that allows users to generate a single shareable link for all uploaded data belonging to them. This simplifies sharing and provides a consolidated view of their contributions.

The alert message for sharing has been updated to accommodate this new master link functionality, displaying a specific message when an empty item name is provided.